fix: ТЗ-11/2 корректировка после ревью
