+++ src/main/java/ru/yandex/practicum/filmorate/storage/GenreDao.java
@@ -8,14 +8,9 @@ import java.util.Optional;
 
 public interface GenreDao {
 
-
-    void createFilmGenre(Film film);
-
     Optional<Genre> getGenreById(int id);
 
     List<Genre> getAllGenres();
-
-    void updateFilmGenre(Film film);
 
     void loadGenres(List<Film> films);
 }
