+++ src/main/resources/application.properties
@@ -1,5 +1,7 @@
 spring.sql.init.mode=always
-spring.datasource.url=jdbc:h2:file:./db/filmorate
+spring.datasource.url=jdbc:h2:file:./db/filmorate;DB_CLOSE_ON_EXIT=FALSE
 spring.datasource.driverClassName=org.h2.Driver
 spring.datasource.username=sa
 spring.datasource.password=password
+spring.jpa.hibernate.ddl-auto=update
+spring.h2.console.enabled=true
